refactor: 사용에 따른 폴더명 변경
checkbox, Input 컴포넌트의 사용처가 한정되어 Auth+@ 로 변경
