+++ src/components/atoms/AuthCheckBox/index.tsx
